Time: 5 ms (14.03%) | Memory: 111.4 MB (43.11%) - LeetSync

+++ 2432-number-of-zero-filled-subarrays/number-of-zero-filled-subarrays.cpp
@@ -1,17 +1,16 @@
 class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
-        long long n = nums.size();
+        int n = nums.size();
         long long nSubArrays = 0; 
 
         long long count = 0; 
-        for(long long i=0; i<n; i++){
+        for(int i=0; i<n; i++){
             if(nums[i] == 0){
                 count++;
             }
             else{
                 nSubArrays += (count * (count +1)) / 2;
-
                 count = 0;
             }
